+++ CHANGELOG.md
@@ -2,9 +2,7 @@
 
 ## Added
 
-## Fixed
-
-## Changed
+- Start adding CLI interface
 
 # 0.1.11 (2025-11-15 / 5b63506)